start splitting up quests block
