+++ client/blocks/quests/index.js
@@ -1,167 +1,29 @@
 import React, {Component} from 'react';
 import formJson from '@quarterto/form-json';
-import {H1, H2} from '../components/heading';
-import Ornamented from '../components/ornamented';
+import {H1, H2} from '../../components/heading';
+import Ornamented from '../../components/ornamented';
 import {withTracker} from 'meteor/react-meteor-data';
-import getCampaignSession from '../../shared/session';
-import {Cards} from '../../shared/collections'
-import idFirst from '../id-first';
+import getCampaignSession from '../../../shared/session';
+import {Cards} from '../../../shared/collections'
+import idFirst from '../../id-first';
 import OdreianDate from 'odreian-date';
 import styled, {keyframes} from 'styled-components';
-import {withCampaign} from '../components/campaign';
-import {background} from '../colors';
-import Portal from 'react-portal';
+import {withCampaign} from '../../components/campaign';
 import {compose} from 'recompose';
 
-const fadeIn = keyframes`
-	0% {
-		opacity: 0;
-		transform: scale(3);
-	}
-
-	90% {
-		opacity: 0.9;
-		transform: scale(0.9);
-	}
-
-	100% {
-		opacity: 1;
-	}
-`;
-
-const fadeOut = keyframes`
-	0%   { opacity: 1; }
-	100% { opacity: 0; }
-`;
-
-const Modal = styled.div`
-	position: fixed;
-	top: 0;
-	left: 0;
-	right: 0;
-	bottom: 0;
-	background: ${background};
-	transform-origin: center;
-	animation-name: ${
-		({animationState}) => ({
-			opening: fadeIn,
-			closing: fadeOut,
-		})[animationState] || 'none'
-	};
-	animation-duration: ${({animationState}) => animationState === 'opening' ? '200ms' : '5s'};
-	animation-fill-mode: forwards;
-	animation-timing-function: linear;
-	animation-iteration-count: 1;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: stretch;
-	text-align: center;
-`;
-
-const QuestHeader = styled.h1`
-	font-family: "Libre Baskerville", sans-serif;
-	font-size: 5em;
-	margin: 0;
-	line-height: 1;
-`;
-
-const ObjectiveHeader = styled.h1`
-	font-family: "Source Sans Pro", sans-serif;
-	font-weight: 300;
-	font-size: 5em;
-	margin: 0;
-	line-height: 1;
-`;
-
-const Splash = ({action, quest, objective, animationState}) => <Modal
-	animationState={animationState}
->
-	{action === 'startQuest' && <ObjectiveHeader>
-		Started:
-	</ObjectiveHeader>}
-	<QuestHeader>
-		<Ornamented ornament='u'>
-			{quest.title}
-		</Ornamented>
-	</QuestHeader>
-	{objective && <ObjectiveHeader>
-		{action === 'completeObjective'
-			? 'Completed: '
-			: ''}
-		{objective.title}
-	</ObjectiveHeader>}
-</Modal>;
-
-const withSplashQuest = withTracker(({campaignId}) => ({
-	splashQuest: getCampaignSession(campaignId).get('splashQuest'),
-}));
-
-const connectQuestSplash = compose(
-	withCampaign,
-	withSplashQuest
-);
-
-class QuestSplash extends Component {
-	state = {
-		splashQuest: null,
-		animationState: 'closed',
-	};
-
-	componentWillReceiveProps(nextProps) {
-		if(nextProps.splashQuest) {
-			this.setState({
-				splashQuest: nextProps.splashQuest,
-				animationState: 'opening',
-			});
-
-			this.timer = setTimeout(() => {
-				this.setState({animationState: 'closing'});
-
-				this.timer = setTimeout(() => {
-					this.setState({
-						splashQuest: null,
-						animationState: 'closed',
-					});
-				}, 5000);
-			}, 10000);
-		} else {
-			this.setState({
-				splashQuest: null,
-				animationState: 'closed',
-			});
-		}
-	}
-
-	render() {
-		if(this.state.splashQuest) {
-			return <Portal isOpened={true}>
-				<Splash
-					{...this.state.splashQuest}
-					animationState={this.state.animationState}
-				/>
-			</Portal>;
-		}
-
-		return null
-	}
-}
-
-const QuestSplashContainer = connectQuestSplash(QuestSplash);
+import QuestSplash from './splash';
 
 const Completed = styled.span`
 	float: right;
 	font-size: 0.7em;
 `;
 
-const getQuestObjectives = ({quest, campaignId}) => Cards.find({
-	type: 'objective',
-	_id: {$in: quest.related || []},
-	campaignId,
-}).fetch();
-
 const withQuestObjectives = withTracker(({quest, campaignId}) => ({
-	objectives: getQuestObjectives({quest, campaignId}),
+	objectives: Cards.find({
+		type: 'objective',
+		_id: {$in: quest.related || []},
+		campaignId,
+	}).fetch(),
 }));
 
 const connectObjectivesList = compose(
@@ -244,7 +106,7 @@ const QuestsList = connectQuestsList(({onCreateQuest, quests, ...props}) => <div
 		<input placeholder='Quest' name='title' />
 		<button>➕</button>
 	</form>}
-	{!onCreateQuest && <QuestSplashContainer />}
+	{!onCreateQuest && <QuestSplash />}
 </div>);
 
 const withQuestControl = withTracker(({campaignId}) => {
